+++ src/adapters/common.rs
@@ -5,3 +5,5 @@ pub mod helpers;
 pub mod keymgmt_functions;
 
 pub mod macros;
+
+pub mod transcoders;
